spotify_search_podcasts/requirements.txt
google-cloud-bigquery==3.5.0
pandas==1.5.3
functions-framework==3.*
numpy==1.23.2
requests==2.28.1
pyarrow==14.0.1
